+++ validator/src/commands/run/args.rs
@@ -236,13 +236,6 @@ pub fn add_args<'a>(app: App<'a, 'a>, default_args: &'a DefaultArgs) -> App<'a, 
             .help("Do not publish the RPC port for use by others"),
     )
     .arg(
-        Arg::with_name("no_port_check")
-            .long("no-port-check")
-            .takes_value(false)
-            .hidden(hidden_unless_forced())
-            .help("Do not perform TCP/UDP reachable port checks at start-up"),
-    )
-    .arg(
         Arg::with_name("account_paths")
             .long("accounts")
             .value_name("PATHS")
@@ -802,9 +795,8 @@ pub fn add_args<'a>(app: App<'a, 'a>, default_args: &'a DefaultArgs) -> App<'a, 
             .takes_value(true)
             .validator(solana_net_utils::is_host)
             .help(
-                "Public IP address for this validator to advertise in gossip. [default: ask \
-                 --entrypoint via its IP echo server, or 127.0.0.1 when --entrypoint is not \
-                 provided]. In a multihoming context (> 1 --bind-address), this flag is ignored \
+                "Public IP address for this validator to advertise in gossip. [default: \
+                 127.0.0.1]. In a multihoming context (> 1 --bind-address), this flag is ignored \
                  and the active --bind-address is advertised instead.",
             ),
     )